refactor(ci): split ci-routing CLI command families (#1174)
Extract content-hash CLI, detect-changes production IO, and shared
flagValue/GITHUB_OUTPUT helpers from the kitchen-sink cli.ts dispatcher.
Add spawn tests so GITHUB_OUTPUT appends for hash-inputs, emit, and
validate-success-marker cannot drop silently.
